rest-api/trusted-auth-server-python/requirements.txt
Flask==2.2.5
Werkzeug==3.0.6
requests==2.32.4
python-dotenv==0.19.0